+++ packages/dev-env/package.json
@@ -1,8 +1,8 @@
 {
   "name": "@atproto/dev-env",
   "version": "0.1.0",
-  "main": "src/test-env.ts",
-  "bin": "dist/cli.js",
+  "main": "src/index.ts",
+  "bin": "dist/bin.js",
   "license": "MIT",
   "repository": {
     "type": "git",
@@ -12,7 +12,7 @@
   "scripts": {
     "build": "node ./build.js",
     "postbuild": "tsc --build tsconfig.build.json",
-    "start": "node dist/cli.js",
+    "start": "node dist/bin.js",
     "prettier": "prettier --check src/",
     "prettier:fix": "prettier --write src/",
     "lint": "eslint . --ext .ts,.tsx",
