+++ packages/dev-env/src/bin.ts
@@ -1,11 +1,7 @@
+import { generateMockSetup } from './mock'
 import { TestNetwork } from './network'
 
 const run = async () => {
-  logStart()
-  const network = await TestNetwork.create()
-}
-
-const logStart = () => {
   console.log(`
 ██████╗
 ██╔═══██╗
@@ -15,4 +11,9 @@ const logStart = () => {
  ╚╝╚═══╝  protocol
 
 [ created by Bluesky ]`)
+
+  const network = await TestNetwork.create()
+  await generateMockSetup(network)
 }
+
+run()
