--- packages/dev-env/src/test-env/index.ts
@@ -1,6 +1,0 @@
-export * from './bsky'
-export * from './network'
-export * from './pds'
-export * from './plc'
-export * from './types'
-export * from './util'
